+++ crates/mako/src/plugins/mod.rs
@@ -2,6 +2,7 @@ pub mod async_runtime;
 pub mod bundless_compiler;
 pub mod context_module;
 pub mod copy;
+pub mod detect_circular_dependence;
 pub mod emotion;
 pub mod graphviz;
 pub mod hmr_runtime;