+++ crates/mako/src/visitors/dynamic_import_to_require.rs
@@ -1,41 +1,85 @@
 use swc_core::common::{Mark, DUMMY_SP};
-use swc_core::ecma::ast::{Expr, ExprOrSpread, Lit};
-use swc_core::ecma::utils::{member_expr, quote_ident, quote_str, ExprFactory};
+use swc_core::ecma::ast::{Expr, ExprOrSpread, Ident, Lit, MemberExpr, Stmt, VarDeclKind};
+use swc_core::ecma::utils::{
+    member_expr, private_ident, quote_ident, quote_str, ExprFactory, IsDirective,
+};
 use swc_core::ecma::visit::{VisitMut, VisitMutWith};
 
 use crate::ast::utils::is_dynamic_import;
-
 pub struct DynamicImportToRequire {
     pub unresolved_mark: Mark,
+    changed: bool,
+    interop: Ident,
 }
 
+impl DynamicImportToRequire {
+    pub fn new(unresolved_mark: Mark) -> Self {
+        let interop = private_ident!("interop");
+        Self {
+            unresolved_mark,
+            changed: false,
+            interop,
+        }
+    }
+}
 // import('xxx') -> Promise.resolve().then(() => require('xxx'))
 impl VisitMut for DynamicImportToRequire {
+    fn visit_mut_module(&mut self, n: &mut swc_core::ecma::ast::Module) {
+        n.visit_mut_children_with(self);
+
+        if self.changed {
+            let insert_at = n
+                .body
+                .iter()
+                .position(|module_item| {
+                    !module_item
+                        .as_stmt()
+                        .map_or(false, |stmt| stmt.is_directive())
+                })
+                .unwrap();
+            let require_interop = quote_ident!("__mako_require__").as_call(
+                DUMMY_SP,
+                vec![quote_str!("@swc/helpers/_/_interop_require_wildcard").as_arg()],
+            );
+
+            let stmt: Stmt = Expr::Member(MemberExpr {
+                span: DUMMY_SP,
+                obj: require_interop.into(),
+                prop: quote_ident!("_").into(),
+            })
+            .into_var_decl(VarDeclKind::Var, self.interop.clone().into())
+            .into();
+            n.body.insert(insert_at, stmt.into());
+        }
+    }
+
     fn visit_mut_expr(&mut self, expr: &mut Expr) {
         if let Expr::Call(call_expr) = expr {
             if is_dynamic_import(call_expr) {
+                self.changed = true;
                 if let ExprOrSpread {
                     expr: box Expr::Lit(Lit::Str(ref mut source)),
                     ..
                 } = &mut call_expr.args[0]
                 {
+                    let source_lazy_require =
+                        quote_ident!(DUMMY_SP.apply_mark(self.unresolved_mark), "require")
+                            .as_call(DUMMY_SP, vec![quote_str!(source.value.clone()).as_arg()])
+                            .into_lazy_arrow(vec![]);
+
                     // Promise.resolve()
                     let promise_resolve: Box<Expr> = member_expr!(DUMMY_SP, Promise.resolve)
                         .as_call(DUMMY_SP, vec![])
                         .into();
 
-                    // () => require( source.value... )
-                    let lazy_require =
-                        quote_ident!(DUMMY_SP.apply_mark(self.unresolved_mark), "require")
-                            .as_call(DUMMY_SP, vec![quote_str!(source.value.clone()).as_arg()])
-                            .into_lazy_arrow(vec![]);
+                    let interop_call = quote_ident!(DUMMY_SP, self.interop.as_ref());
+                    let promised_require: Expr = member_expr!(@EXT,DUMMY_SP, promise_resolve, then)
+                        .as_call(DUMMY_SP, vec![source_lazy_require.as_arg()]);
+                    let interop_require =
+                        member_expr!(@EXT, DUMMY_SP, promised_require.into(), then)
+                            .as_call(DUMMY_SP, vec![interop_call.as_arg()]);
 
-                    // Promise.resolve().then(() => require("xxx"))
-                    let promised_lazy_require: Expr =
-                        member_expr!(@EXT,DUMMY_SP, promise_resolve, then)
-                            .as_call(DUMMY_SP, vec![lazy_require.as_arg()]);
-
-                    *expr = promised_lazy_require;
+                    *expr = interop_require;
                 }
             }
         }
@@ -55,7 +99,9 @@ mod tests {
     fn test_basic_import() {
         assert_eq!(
             run(r#"const testModule = import('test-module');"#,),
-            r#"const testModule = Promise.resolve().then(()=>require("test-module"));"#.trim()
+            r#"var interop = __mako_require__("@swc/helpers/_/_interop_require_wildcard")._;
+const testModule = Promise.resolve().then(()=>require("test-module")).then(interop);"#
+                .trim()
         );
     }
 
@@ -63,7 +109,8 @@ mod tests {
     fn test_chained_import() {
         assert_eq!(
             run(r#"import('test-module').then(() => (import('test-module-2')));"#,),
-            r#"Promise.resolve().then(()=>require("test-module")).then(()=>(Promise.resolve().then(()=>require("test-module-2"))));"#.trim()
+            r#"var interop = __mako_require__("@swc/helpers/_/_interop_require_wildcard")._;
+Promise.resolve().then(()=>require("test-module")).then(interop).then(()=>(Promise.resolve().then(()=>require("test-module-2")).then(interop)));"#.trim()
         );
         assert_eq!(
             run(r#"
@@ -74,10 +121,11 @@ Promise.all([
 ]).then(() => {});
             "#,),
             r#"
+var interop = __mako_require__("@swc/helpers/_/_interop_require_wildcard")._;
 Promise.all([
-    Promise.resolve().then(()=>require("test-1")),
-    Promise.resolve().then(()=>require("test-2")),
-    Promise.resolve().then(()=>require("test-3"))
+    Promise.resolve().then(()=>require("test-1")).then(interop),
+    Promise.resolve().then(()=>require("test-2")).then(interop),
+    Promise.resolve().then(()=>require("test-3")).then(interop)
 ]).then(()=>{});
             "#
             .trim()
@@ -92,8 +140,9 @@ import(/* test comment */ 'my-module');
 import('my-module' /* test comment */ );
             "#,),
             r#"
-Promise.resolve().then(()=>require("my-module"));
-Promise.resolve().then(()=>require("my-module"));
+var interop = __mako_require__("@swc/helpers/_/_interop_require_wildcard")._;
+Promise.resolve().then(()=>require("my-module")).then(interop);
+Promise.resolve().then(()=>require("my-module")).then(interop);
             "#
             .trim()
         );
@@ -128,9 +177,7 @@ import(`3-${MODULE}`);
         let mut test_utils = TestUtils::gen_js_ast(js_code);
         let ast = test_utils.ast.js_mut();
         GLOBALS.set(&test_utils.context.meta.script.globals, || {
-            let mut visitor = DynamicImportToRequire {
-                unresolved_mark: ast.unresolved_mark,
-            };
+            let mut visitor = DynamicImportToRequire::new(ast.unresolved_mark);
             ast.ast.visit_mut_with(&mut visitor);
         });
         test_utils.js_ast_to_code()