+++ e2e/fixtures/config.dynamic_import_to_require/src/lazy.ts
@@ -1,0 +1,3 @@
+export default {
+  lazy: 'lazy'
+}