+++ e2e/fixtures/config.public_path.with-assignment/src/index.tsx
@@ -1,0 +1,3 @@
+console.log(1);
+__webpack_public_path__ = '/foo/';
+__mako_public_path__ = '/bar/';